Combined both tor testinfra common tests together
